+++ packages/server/src/adapters/express.ts
@@ -1,11 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import type * as express from 'express';
-import {
-  BaseOptions,
-  CreateContextFn,
-  CreateContextFnOptions,
-  requestHandler,
-} from '../http';
+import { BaseOptions, CreateContextFn, CreateContextFnOptions } from '../http';
+import { requestHandler } from '../http';
 import { AnyRouter } from '../router';
 
 export type CreateExpressContextOptions = CreateContextFnOptions<
